Add force-push diff checking (#56)
Move force-push comment into background thread and also add diff
check of the before/after commit hashes as compared to their
respective fork points from the base branch.

Needed to be a bit creative with the force push diff so that octobot could 
fetch it, since non-referenced commits are not fetched. So, we create a 
temporary branch for the before hash, fetch, then delete that branch.

Also remove 'Compare' URL link since it is kinda misleading and
doesn't really diff the two commits meaningfully, imo.