Add test for finding base commit and bring back --fork-point
I think "--fork-point" is actually what we're looking for, and should
really only ever fail in the case that the branch has been rewritten,
which shouldn't happen for master or release branches. And this time,
we have a fallback to regular merge-base for safety.

Note: This test also reproduced the failure with --is-ancestor,
which I had used in find_base_branch_commit originally.

Also replace `unwrap` calls with better assertions for better feedback