Use full commit hash in 'cherry-picked from'. (#81)
If a repository gets large, then it's possible that
there will be collisions on the short hash. The
pull-request number should disambiguate it, but it'd
be cool if the hash also did not collide.

GitHub will shorten full hashes to the short form in
the pull-request UI, so I think it should still appear
the same.