Add explicit JIRA project list to repo config
Now that we are going to be adding versions and not just comments,
we want to be a bit more sure that the jira project is the right one
for that repo.

This also paves the way for allowing separate JIRA projects for
different release branches in the case of the monolothic repo
scenario.