Fix void HTTP responses
There are times when we want to ignore the response from the server,
whether because we don't care about it, or because it is not
JSON, as it turns out is the case for slack incoming webhooks.

I wanted to do this w/ specialization on VoidResponse, but it looks
like that is not in rust stable yet. There is probably a fancier
way to do this, but this works. :)