Drop iron and use hyper directly (#124)
- Iron was dependening on an older non-async version of hyper,
  and since upgrading http client code to latest hyper, I didn't
  want octobot to have two versions.
- There was a issue in iron (issue 444) I ran into in the past
  that wasn't going to be fixed until async hyper.
- Really, iron was more than we needed. Octobot wasn't really using
  iron for much except routing and before filters, which were
  pretty easy to do directly against hyper's api.
- This might let us support SSL directly in octobot

One change here that had to be made was for the github hook verifier.
Because it needed to consume the Request object to read the entire
body, I moved that check into the github handler. It would have
over complicated things to try to make it fit as a filter, and
there's only one handler, so no need to.