Move user config to sqlite (#157)
This removes the need for the initial creation of the users.json and added
in the github host into it, as well as moves to a more structured format.

This addresses the first half of #153 as well as #139 by doing operations
on a single user at a time from the admin page.

Note: This also removes the multi-tenancy of allowing users to be tied 
to different github hosts. This idea was never really completed, and 
ended up being an unnecessary complication.