Updating libraries to latest versions
Whew. Looks like tokio and hyper had quite a big API change in the
latest version.

No longer pass around a reactor Remote to construct a HTTPClient.
Instead it uses `tokio::spawn` to submit tasks to the context-relevant
event loop. One offshoot of this was that `server::main::start`
needed to be run inside the event loop.