Strip out previous merge title prefixes
If a change was backported multiple times, then it would end up
with a rather unsightly PR title. This was noticed in particular
with the use of `-next` branches (see #229).

Since backported PR's are already annotated by octobot with the
pull request number, the history of where it was originally
backported from is already maintained. The title should just
indicate the source of this particular backport.