Add support for `-next` release branches (#229)
While a release branch is stabilizing, e.g. `release/1.1`, it is
helpful to be able to continue to backport fixes to a `release/1.1-next`
staging release branch.

Generally, here is how the pattern works:
  * Ship release/1.1
  * Branch release/1.1-next from release/1.1
  * Continue to backport important but non-critical changes on release/1.1-next
  * Backport all critical changes directly to release/1.1
  * When release/1.1 is happy, merge release/1.1-next -> release/1.1
  * Repeat.

Benefits:
  * allows changes destined for a future release to continue to get merged in
  * allows merge conflicts to get resolved now rather than
    postponed/stacked/forgotten
  * avoids just leaving lots of PRs open
  * avoids a huge merge party

Specifically, this PR lets octobot know about these `-next` branches so
that versions and JIRAs are not updated when code is merged to avoid
wrong and potentially confusing version information in JIRA.

Note that octobot will still submit "PR submitted on branch ..." comments
to provide some kind of visibility/tracking in JIRA that some
backporting has been done.