Reword backport section of README
> It will auto-cherry-pick this PR after it is merged to "release/1.0"

The original wording sounded like "after it is merged to release/1.0, it will cherry pick to release/1.0", which doesn't make sense. I _think_ it is trying to say that after merging the original PR (to "master" or whatever it targets), it will cherry pick to release/1.0. I tried to make this a little more clear.