Treat empty version script output as a failure (#246)
There are some version script failures that do not cause firejail
to return a non-zero exit code. The case we found appeared to be
due to trying to read from stdin(?), which conflicted with the
`--private-dev` option.

Furthermore, version scripts should never really return empty strings.
If there is a version script, it should always return a version.

So, we now treat empty versions as version script failures.