Add optional login_suffix config for JIRA.
If a login_suffix field is provided in the JIRA config, it will be
appended to the logged-in username to pre-fill the username field in the
JIRA auth dialog.

For example, if for every octobot user "$user", the corresponding JIRA
username is "$user@company.com", setting jira.login_suffix =
"@company.com" will mean users don't have to remember to adjust their
username in the login prompt every time they do a pending version merge.