Don't add pending version to "See" references. (#311)
The "See" keyword allows you to mention a JIRA in a commit message
without transitioning it. Such references should not have a pending
version applied, as they are not intended to fix the issue. If a pending
version is applied, the mentioned JIRA will get a fix-version when the
next rollup is performed.

Co-authored-by: Matt Hauck <matt.hauck@tanium.com>