Supply Github commit/PR url for slack thread conversation
Need to supply the URL to the thing we are talking about, otherwise we would have to search through the message and the attachments for what the URL.  Then if there are multiple URLs which one do you choose.  This type of logic shouldn't be the job of the Slack messenger to determine.