feat: allow threading of messages (#324)
Adds a new option to repo to allow for threading of GitHub PR content within Slack channels. The option only applies to public channels.

If the previous message is found within a local DB of slack threads, then the message will be sent with the parent thread's ID as the value within SlackMessage.ts.
If the previous message cannot be found, then the message will be sent to the channel.