fix: remove empty string when loading `muted_repos`
Users that didn't have any muted repos would load an empty string, which
when split would give an array containing that empty string.

This was causing test failures that created Repos with no name, which
defaulted ot an empty string and then mateched that empty string,
causing messages the test expected to send not to be sent.