feat: add team members to participants (#344)
This will send out slack notifications to team members when the team is added as an assignee on a PR. Note, when one of the team member reviews the PR, the team is removed from the PR. This means that the team member will only receive the initial PR request and no updates after the team is removed. Receiving the initial PR request slack message is still very helpful.