Change teams lookup to use team id
Using the members_url and removing `{/members}` wasn't working, and
kinda relied on a contract that didn't seem guaranteed.

Changed up the team member lookup to use (org_id, team_id). We didn't
do this before to avoid yet another lookup for the team org, but then
realized that teams must belong to the organization that the PR belongs to.
So, we already have the org id.