Update to rust 1.84.
This is the latest stable rust toolchain.
I had to run `cargo update` as some of the older
deps did not support latest rust.

I noticed a dependabot PR failed regarding a rust version mismatch:
https://github.com/tanium/octobot/pull/391

I'm not sure this will fix it, but it seems like a good step anyways.