Support Jira auth via token. (#403)
Allow API request to provide a jira token instead of basic auth.
This allows CLI automation to use less sensitive credentials when calling the API.
The use of the /session endpoint to verify the credentials is a bit awkward --
basic auth must be provided as parameters but a token must be provided as a header.

An invalid token header does not dump a HTML document, I just get

```
{"errorMessages":["You are not authenticated. Authentication required to perform this operation."],"errors":{}}
```

So I don't think the same concerns around invalid basic auth apply.

I verified auth with a temporary test like so:

```
\#[tokio::test]
async fn test_rrc() {
    let config = &JiraConfig {
        host: "jira.corp.tanium.com".into(),
        auth: crate::config::JiraAuth::Token("my-token".into()),
        progress_states: None,
        review_states: None,
        resolved_states: None,
        fixed_resolutions: None,
        fix_versions_field: None,
        pending_versions_field: None,
        restrict_comment_visibility_to_role: None,
        login_suffix: None,
    };
    let s = JiraSession::new(config, None).await.unwrap();
    let i = s.get_issue("SWM-123").await.unwrap();
    eprintln!("{:?}", i.key)
}
```