Add `frozen_states` jira config to block transitions. (#404)
If an issue is in a resolved state and a PR is opened, octobot will attempt to transition it to a review state.
This is generally not desireable, and in particular means backports (even those automated by octobot) will un-resolve issues.
In some cases, the Jira configuration might disallow these transitions, but we don't want to rely on that in octobot.
Add a config for "frozen states". Octobot will not attempt to transition issues once they reach one of these states.

An alterative I considered is disabling transitions on release branches.
However, I thought it was possible you might have a ticket go _only_ into a release branch and still want it resolved.